requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-extra-fields==3.0.2
djangorestframework==3.12.4
Pillow==9.1.1
psycopg2==2.9.2
python-dotenv==0.20.0
pytz==2021.3
sqlparse==0.4.2
